Refactor AB flag storage and handling for testWorkerLogic
