+++ src/domain/ab/testWorkerLogic/index.ts
@@ -1,10 +1,10 @@
 import { getIsFlagEnabled } from "config/ab";
 
-import { buildMarketsValuesRequest as buildMarketsValuesRequestEmbedded } from "domain/ab/testWorkerLogic/withEmbeddedMulticallHashData/buildMarketsValuesRequest";
-import { buildMarketsConfigsRequest as buildMarketsConfigsRequestEmbedded } from "domain/ab/testWorkerLogic/withEmbeddedMulticallHashData/buildMarketsConfigsRequest";
+import { buildMarketsValuesRequest as buildMarketsValuesRequestEmbedded } from "domain/ab/testWorkerLogic/enabled/buildMarketsValuesRequest";
+import { buildMarketsConfigsRequest as buildMarketsConfigsRequestEmbedded } from "domain/ab/testWorkerLogic/enabled/buildMarketsConfigsRequest";
 
-import { buildMarketsValuesRequest as buildMarketsValuesRequestAsync } from "domain/ab/testWorkerLogic/withAsyncHashData/buildMarketsValuesRequest";
-import { buildMarketsConfigsRequest as buildMarketsConfigsRequestAsync } from "domain/ab/testWorkerLogic/withAsyncHashData/buildMarketsConfigsRequest";
+import { buildMarketsValuesRequest as buildMarketsValuesRequestAsync } from "domain/ab/testWorkerLogic/disabled/buildMarketsValuesRequest";
+import { buildMarketsConfigsRequest as buildMarketsConfigsRequestAsync } from "domain/ab/testWorkerLogic/disabled/buildMarketsConfigsRequest";
 
 export const buildMarketsValuesRequest = getIsFlagEnabled("testWorkerLogic")
   ? buildMarketsValuesRequestEmbedded
